+++ test/support/mocks.ex
@@ -1,0 +1,5 @@
+Mox.defmock(Caint.Deepl.MockApiImpl, for: Caint.Deepl.Api)
+
+defmodule Caint.Mocks do
+  @moduledoc false
+end
