+++ test/test_helper.exs
@@ -1,1 +1,2 @@
+Application.put_env(:caint, :deepl_api_impl, Caint.Deepl.MockApiImpl)
 ExUnit.start()
